baseline-apps/image-convertor/requirements.txt
Pillow
boto3
aws-lambda-powertools[tracer]  # Tracer requires AWS X-Ray SDK dependency
